third_party/blink/web_tests/platform/win/external/wpt/css/css-overscroll-behavior/overscroll-behavior-root-expected.txt

This is a testharness.js-based test.
[FAIL] overscroll-behavior in root works
  assert_equals: overscroll-behavior should work on the root expected 775 but got 675
Harness: the test ran to completion.
